services/auth/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
python-dotenv==1.0.1
aws-wsgi==0.2.7
PyMySQL==1.1.1